# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "query",
        "parameters": [
          {
            "name": "phrases",
            "in": "query",
            "description": "Comma-separated phrases to query (max 10)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Start date (YYYY-MM-DD). Default: 2011-01-01",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "End date (YYYY-MM-DD). Default: today",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/totals": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "totals",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "Start date (YYYY-MM-DD). Default: 2011-01-01",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "End date (YYYY-MM-DD). Default: today",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket totals for normalization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "version",
          "tokenizer_version"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "tokenizer_version": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "Point": {
        "type": "object",
        "required": [
          "t",
          "v"
        ],
        "properties": {
          "t": {
            "type": "string",
            "description": "Bucket timestamp (YYYY-MM-DD)"
          },
          "v": {
            "type": "integer",
            "format": "int32",
            "description": "Raw occurrence count (client computes relative frequency using /totals)",
            "minimum": 0
          }
        }
      },
      "QueryMeta": {
        "type": "object",
        "required": [
          "tokenizer_version",
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "tokenizer_version": {
            "type": "string"
          }
        }
      },
      "QueryParams": {
        "type": "object",
        "required": [
          "phrases"
        ],
        "properties": {
          "end": {
            "type": "string",
            "description": "End date (YYYY-MM-DD). Default: today",
            "nullable": true
          },
          "phrases": {
            "type": "string",
            "description": "Comma-separated phrases to query (max 10)"
          },
          "start": {
            "type": "string",
            "description": "Start date (YYYY-MM-DD). Default: 2011-01-01",
            "nullable": true
          }
        }
      },
      "QueryResponse": {
        "type": "object",
        "required": [
          "series",
          "meta"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/QueryMeta"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Series"
            }
          }
        }
      },
      "Series": {
        "type": "object",
        "required": [
          "phrase",
          "normalized",
          "status",
          "points"
        ],
        "properties": {
          "normalized": {
            "type": "string",
            "description": "Normalized/tokenized form used for lookup"
          },
          "phrase": {
            "type": "string",
            "description": "Original phrase from input"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "description": "Data points (empty for NotIndexed/Invalid)"
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatus"
          }
        }
      },
      "SeriesStatus": {
        "type": "string",
        "enum": [
          "indexed",
          "not_indexed",
          "invalid"
        ]
      },
      "TotalPoint": {
        "type": "object",
        "required": [
          "t",
          "v"
        ],
        "properties": {
          "t": {
            "type": "string",
            "description": "Bucket timestamp (YYYY-MM-DD)"
          },
          "v": {
            "type": "integer",
            "format": "int64",
            "description": "Total n-gram count for this day and order",
            "minimum": 0
          }
        }
      },
      "TotalSeries": {
        "type": "object",
        "required": [
          "n",
          "points"
        ],
        "properties": {
          "n": {
            "type": "integer",
            "format": "int32",
            "description": "N-gram order (1, 2, or 3)",
            "minimum": 0
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TotalPoint"
            },
            "description": "Sparse daily total counts (only non-zero days)"
          }
        }
      },
      "TotalsMeta": {
        "type": "object",
        "required": [
          "tokenizer_version",
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "tokenizer_version": {
            "type": "string"
          }
        }
      },
      "TotalsParams": {
        "type": "object",
        "properties": {
          "end": {
            "type": "string",
            "description": "End date (YYYY-MM-DD). Default: today",
            "nullable": true
          },
          "start": {
            "type": "string",
            "description": "Start date (YYYY-MM-DD). Default: 2011-01-01",
            "nullable": true
          }
        }
      },
      "TotalsResponse": {
        "type": "object",
        "required": [
          "totals",
          "meta"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/TotalsMeta"
          },
          "totals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TotalSeries"
            },
            "description": "One entry per n-gram order (1, 2, 3), each containing sparse daily totals"
          }
        }
      }
    }
  }
}